Add base content for MbedTLS DTLS support
This work is based on LukasKarel work https://github.com/eclipse/wakaama/pull/626
This commit contains a new structure for the connection layer which
is more flexible and decouples layers in a better way.

DTLS is not working properly. For now, messages can be sent, but
packet decryption on client (maybe also on server) fails.